chore: Split useNetworkState into discrete stores
