+++ package.json
@@ -49,6 +49,7 @@
     "@types/react-dom": "^18.0.11",
     "@vitejs/plugin-react": "^3.1.0",
     "electron": "^23.2.0",
+    "eslint-import-resolver-typescript": "^3.5.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "typescript": "^5.0.2",
