+++ src/App.tsx
@@ -40,5 +40,3 @@ export const App = () => {
     </>
   );
 };
-
-export default App;
