+++ src/components/canvas/atoms/hackBotModel/HackBotModel.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useRef } from 'react';
+import { useEffect, useRef } from 'react';
 import * as THREE from 'three';
 import { Group } from 'three';
 import { useFrame } from '@react-three/fiber';
