+++ src/components/canvas/atoms/networkOrb/NetworkOrb.tsx
@@ -1,4 +1,4 @@
-import React, { useRef } from 'react';
+import { useRef } from 'react';
 import { Group } from 'three';
 import { useControls } from 'leva';
