+++ src/components/canvas/molecules/vertexCollection/VertexCollection.tsx
@@ -1,23 +1,23 @@
-import React from 'react';
 import { VertexModel } from '../../atoms/vertexModel/VertexModel';
-import { HackBotProps, Vertex } from '../../../controllers/networkController/types';
+import { HackBotProps } from '@/store/hackBot/types';
+import { Vertex } from '@/store/vertex/types';
 
 interface Props {
   createHackBot: ({ vertex }: HackBotProps) => void,
-  removeHackBot: (uuid: string) => void,
+  deleteHackBot: (uuid: string) => void,
   vertices: Vertex[],
 }
 
-export const VertexCollection = ({ createHackBot, removeHackBot, vertices }: Props) => {
+export const VertexCollection = ({ createHackBot, deleteHackBot, vertices }: Props) => {
   return (
     <>
        {
          vertices.map((vertex, i) => {
            return (
              <VertexModel
-               key={`Vertex ${i}: ${vertex.vector}`}
+               key={`Vertex ${i}: ${vertex.vector.x}`}
                createHackBot={createHackBot}
-               removeHackBot={removeHackBot}
+               deleteHackBot={deleteHackBot}
                vertex={vertex}
                uuid={vertex.uuid}
              />
