+++ src/components/canvas/scenes/mainMenu/MainMenu.tsx
@@ -1,7 +1,8 @@
-import React, { useState } from 'react';
-import useScene, { SCENE } from '../../../managers/sceneManager/useSceneState';
+import { useState } from 'react';
+import useScene from '@/store/scene/useSceneState';
 import { Text } from '@react-three/drei';
 import * as THREE from 'three';
+import { SCENE } from '@/store/scene/types';
 
 export const MainMenu = () => {
   const { updateScene } = useScene();
