+++ src/components/canvas/scenes/match/Match.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { NetworkController } from '../../../controllers/networkController/NetworkController';
 
 export const Match = () => {
