+++ src/store/edge/types.ts
@@ -1,0 +1,6 @@
+export interface EdgeState {
+  maxEdgeLengthPercentage: number,
+
+  // Actions
+  updateMaxEdgeLengthPercentage: (newMaxEdgeLengthPercentage: number) => void,
+}
