+++ src/store/edge/useEdgeState.ts
@@ -1,0 +1,17 @@
+import { create } from 'zustand';
+import { EdgeState } from '@/store/edge/types';
+
+export default create<EdgeState>((set, get) => {
+  return {
+    maxEdgeLengthPercentage: 0.97,
+
+    // Actions
+    updateMaxEdgeLengthPercentage: (newMaxEdgeLengthPercentage: number) => {
+      set(() => {
+        return {
+          maxEdgeLengthPercentage: newMaxEdgeLengthPercentage,
+        };
+      });
+    },
+  };
+});
