+++ src/store/network/types.ts
@@ -1,0 +1,10 @@
+export interface NetworkState {
+  orbOpacity: number,
+  orbRadius: number,
+  radius: number,
+
+  // Actions
+  createNetwork: () => void,
+  updateOrbOpacity: (newOpacity: number) => void,
+  updateOrbRadius: (newRadius: number) => void,
+}
